pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalyst-pack-schemas"
version = "1.0.0"
description = "Shared schemas and validation for Catalyst Knowledge Packs"
readme = "README.md"
authors = [
    {name = "Catalyst Team", email = "team@catalyst.dev"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "requests>=2.25.0",
    "pathlib>=1.0.0; python_version<'3.4'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/catalyst-pack-schemas"
Repository = "https://github.com/your-org/catalyst-pack-schemas"
Issues = "https://github.com/your-org/catalyst-pack-schemas/issues"

[project.scripts]
catalyst-packs = "catalyst_pack_schemas.cli:main"

[tool.setuptools]
packages = ["catalyst_pack_schemas"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100